Week 7 HW Submission
(same as other submission, but with a description... forgot) Improved my box model and cleaned up most of my margins. Also generally polished the site and transformed the index page into a neat welcome screen. I'm having a seemingly anomalous problem with the width of the gallery page not conforming to the rest, and would appreciate a look at it.