Week 7 Homework Submission
Minor aesthetic and functionality tune-ups mostly this week. I added an aside for the gallery, to jump around in-page. Tried out a floated picture in my bio, but am having an overflow problem that isn't fixed by setting overflow: auto. I still have an issue with the scale of the gallery page being inexplicably different than the others. That's it.