Final Project Repo Submission
This is my final iteration of benrodia.com for the purposes of IDM221. The main attraction is the gallery, which is still not fully fleshed out (I have a few videos to add), but there are no more filler boxes, so everything is as it would appear online, if less full. What a sentence. I'm overall happy with the amount of features and code I've been able to incorporate, except for scalability, which I'm sure will be covered in depth next class.